+++ relay-event-normalization/src/eap/ai.rs
@@ -1,6 +1,6 @@
 use std::time::Duration;
 
-use relay_conventions::consts::*;
+use relay_conventions::attributes::*;
 use relay_event_schema::protocol::Attributes;
 use relay_protocol::Annotated;