+++ relay-event-normalization/src/eap/mod.rs
@@ -7,7 +7,7 @@ use std::net::IpAddr;
 
 use chrono::{DateTime, Utc};
 use relay_common::time::UnixTimestamp;
-use relay_conventions::consts::*;
+use relay_conventions::attributes::*;
 use relay_conventions::{AttributeInfo, WriteBehavior};
 use relay_event_schema::protocol::{AttributeType, Attributes, BrowserContext, Geo};
 use relay_protocol::{Annotated, ErrorKind, Meta, Remark, RemarkType, Value};
@@ -347,7 +347,15 @@ pub fn normalize_user_geo(
 }
 
 /// Normalizes the [DSC](DynamicSamplingContext) into [`Attributes`].
-pub fn normalize_dsc(attributes: &mut Annotated<Attributes>, dsc: Option<&DynamicSamplingContext>) {
+///
+/// If `is_segment` is set to `false`, the function will only add select attributes that are
+/// necessary on every span - both segment and non-segment - for dynamic sampling to work. More
+/// attributes are added when `is_segment` is set to `true`.
+pub fn normalize_dsc(
+    attributes: &mut Annotated<Attributes>,
+    is_segment: &Annotated<bool>,
+    dsc: Option<&DynamicSamplingContext>,
+) {
     let Some(dsc) = dsc else { return };
 
     let attributes = attributes.get_or_insert_with(Default::default);
@@ -356,23 +364,26 @@ pub fn normalize_dsc(attributes: &mut Annotated<Attributes>, dsc: Option<&Dynami
     if attributes.contains_key(SENTRY__DSC__TRACE_ID) {
         return;
     }
-
     attributes.insert(SENTRY__DSC__TRACE_ID, dsc.trace_id.to_string());
-    attributes.insert(SENTRY__DSC__PUBLIC_KEY, dsc.public_key.to_string());
-    if let Some(release) = &dsc.release {
-        attributes.insert(SENTRY__DSC__RELEASE, release.clone());
-    }
-    if let Some(environment) = &dsc.environment {
-        attributes.insert(SENTRY__DSC__ENVIRONMENT, environment.clone());
-    }
+
     if let Some(transaction) = &dsc.transaction {
         attributes.insert(SENTRY__DSC__TRANSACTION, transaction.clone());
     }
-    if let Some(sample_rate) = dsc.sample_rate {
-        attributes.insert(SENTRY__DSC__SAMPLE_RATE, sample_rate);
-    }
-    if let Some(sampled) = dsc.sampled {
-        attributes.insert(SENTRY__DSC__SAMPLED, sampled);
+
+    if is_segment.value().is_some_and(|is_segment| *is_segment) {
+        attributes.insert(SENTRY__DSC__PUBLIC_KEY, dsc.public_key.to_string());
+        if let Some(release) = &dsc.release {
+            attributes.insert(SENTRY__DSC__RELEASE, release.clone());
+        }
+        if let Some(environment) = &dsc.environment {
+            attributes.insert(SENTRY__DSC__ENVIRONMENT, environment.clone());
+        }
+        if let Some(sample_rate) = dsc.sample_rate {
+            attributes.insert(SENTRY__DSC__SAMPLE_RATE, sample_rate);
+        }
+        if let Some(sampled) = dsc.sampled {
+            attributes.insert(SENTRY__DSC__SAMPLED, sampled);
+        }
     }
 }
 
@@ -665,20 +676,22 @@ pub fn write_legacy_attributes(attributes: &mut Annotated<Attributes>) {
     };
 
     // Map of new sentry conventions attributes to legacy SpanV1 attributes
+    #[allow(
+        deprecated,
+        reason = "Writing possibly deprecated legacy attributes is the point of this function."
+    )]
     let current_to_legacy_attributes = [
         // DB attributes
         (DB__QUERY__TEXT, SENTRY__DESCRIPTION),
         (SENTRY__NORMALIZED_DB_QUERY, SENTRY__NORMALIZED_DESCRIPTION),
         (DB__OPERATION__NAME, SENTRY__ACTION),
-        #[allow(
-            deprecated,
-            reason = "Writing possibly deprecated legacy attributes is the point of this function."
-        )]
         (DB__SYSTEM__NAME, DB__SYSTEM),
         // HTTP attributes
         (SERVER__ADDRESS, SENTRY__DOMAIN),
         (HTTP__REQUEST__METHOD, SENTRY__ACTION),
         (HTTP__RESPONSE__STATUS_CODE, SENTRY__STATUS_CODE),
+        // Transaction
+        (SENTRY__SEGMENT__NAME, SENTRY__TRANSACTION),
     ];
 
     for (current_attribute, legacy_attribute) in current_to_legacy_attributes {
@@ -717,9 +730,88 @@ pub fn write_legacy_attributes(attributes: &mut Annotated<Attributes>) {
 
 #[cfg(test)]
 mod tests {
-    use relay_protocol::{Empty, SerializableAnnotated};
+    use relay_protocol::{Empty, SerializableAnnotated, assert_annotated_snapshot};
 
     use super::*;
+
+    fn mock_dsc(transaction: Option<&str>) -> DynamicSamplingContext {
+        DynamicSamplingContext {
+            trace_id: "67e5504410b1426f9247bb680e5fe0c8".parse().unwrap(),
+            public_key: "12345678901234567890123456789012".parse().unwrap(),
+            release: None,
+            environment: None,
+            transaction: transaction.map(str::to_owned),
+            sample_rate: None,
+            user: Default::default(),
+            replay_id: None,
+            sampled: None,
+            other: Default::default(),
+        }
+    }
+
+    #[test]
+    fn test_normalize_dsc_child_span_no_dsc() {
+        let mut attributes = Annotated::empty();
+        normalize_dsc(&mut attributes, &Annotated::new(false), None);
+        assert!(attributes.value().is_none());
+    }
+
+    #[test]
+    fn test_normalize_dsc_child_span_no_transaction() {
+        let mut attributes = Annotated::empty();
+        let dsc = mock_dsc(None);
+        normalize_dsc(&mut attributes, &Annotated::new(false), Some(&dsc));
+        assert_annotated_snapshot!(attributes, @r#"
+        {
+          "sentry.dsc.trace_id": {
+            "type": "string",
+            "value": "67e5504410b1426f9247bb680e5fe0c8"
+          }
+        }
+        "#);
+    }
+
+    #[test]
+    fn test_normalize_dsc_child_span() {
+        let mut attributes = Annotated::empty();
+        let dsc = mock_dsc(Some("/some/endpoint"));
+        normalize_dsc(&mut attributes, &Annotated::new(false), Some(&dsc));
+        assert_annotated_snapshot!(attributes, @r#"
+        {
+          "sentry.dsc.trace_id": {
+            "type": "string",
+            "value": "67e5504410b1426f9247bb680e5fe0c8"
+          },
+          "sentry.dsc.transaction": {
+            "type": "string",
+            "value": "/some/endpoint"
+          }
+        }
+        "#);
+    }
+
+    #[test]
+    fn test_normalize_dsc_segment() {
+        let mut attributes = Annotated::empty();
+        let dsc = mock_dsc(Some("/some/endpoint"));
+        normalize_dsc(&mut attributes, &Annotated::new(true), Some(&dsc));
+        assert_annotated_snapshot!(attributes, @r#"
+        {
+          "sentry.dsc.public_key": {
+            "type": "string",
+            "value": "12345678901234567890123456789012"
+          },
+          "sentry.dsc.trace_id": {
+            "type": "string",
+            "value": "67e5504410b1426f9247bb680e5fe0c8"
+          },
+          "sentry.dsc.transaction": {
+            "type": "string",
+            "value": "/some/endpoint"
+          }
+        }
+        "#);
+    }
 
     #[test]
     fn test_normalize_received_none() {
@@ -730,7 +822,7 @@ mod tests {
             DateTime::from_timestamp_nanos(1_234_201_337),
         );
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.observed_timestamp_nanos": {
             "type": "string",
@@ -757,7 +849,7 @@ mod tests {
             DateTime::from_timestamp_nanos(1_234_201_337),
         );
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r###"
+        assert_annotated_snapshot!(attributes, @r###"
         {
           "sentry.observed_timestamp_nanos": {
             "type": "string",
@@ -854,7 +946,7 @@ mod tests {
         let mut attributes = Annotated::<Attributes>::from_json(json).unwrap();
         normalize_attribute_types(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "double_with_i64": {
             "type": "double",
@@ -1026,7 +1118,7 @@ mod tests {
             }),
         );
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "browser.name": {
             "type": "string",
@@ -1066,7 +1158,7 @@ mod tests {
             }),
         );
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "browser.name": {
             "type": "string",
@@ -1077,7 +1169,7 @@ mod tests {
             "value": "13.3.7"
           }
         }
-        "#,
+        "#
         );
     }
 
@@ -1103,7 +1195,7 @@ mod tests {
             })
         });
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -1143,7 +1235,7 @@ mod tests {
 
         normalize_user_geo(&mut attributes, |_| unreachable!());
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -1154,7 +1246,7 @@ mod tests {
             "value": "Foo Hausen"
           }
         }
-        "#,
+        "#
         );
     }
 
@@ -1217,7 +1309,7 @@ mod tests {
 
         normalize_attribute_names_inner(&mut attributes, mock_attribute_info);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r###"
+        assert_annotated_snapshot!(attributes, @r###"
         {
           "backfill.empty": {
             "type": "string",
@@ -1290,7 +1382,7 @@ mod tests {
 
         normalize_sentry_op(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "db.operation.name": {
             "type": "string",
@@ -1336,7 +1428,7 @@ mod tests {
 
         normalize_db_attributes(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "db.operation.name": {
             "type": "string",
@@ -1402,7 +1494,7 @@ mod tests {
 
         normalize_db_attributes(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "db.collection.name": {
             "type": "string",
@@ -1528,7 +1620,7 @@ mod tests {
 
         normalize_db_attributes(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -1574,7 +1666,7 @@ mod tests {
 
         normalize_http_attributes(&mut attributes, &[]);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -1632,7 +1724,7 @@ mod tests {
 
         normalize_http_attributes(&mut attributes, &[]);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -1693,7 +1785,7 @@ mod tests {
             &["application.www.xn--85x722f.xn--55qx5d.cn".to_owned()],
         );
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -1751,7 +1843,7 @@ mod tests {
 
         normalize_db_attributes(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "db.collection.name": {
             "type": "string",
@@ -1810,7 +1902,7 @@ mod tests {
         normalize_attribute_names(&mut attributes);
         normalize_http_attributes(&mut attributes, &[]);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -1856,7 +1948,7 @@ mod tests {
 
         normalize_http_attributes(&mut attributes, &[]);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -1926,7 +2018,7 @@ mod tests {
 
         write_legacy_attributes(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "db.collection.name": {
             "type": "string",
@@ -1999,7 +2091,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.category": {
             "type": "string",
@@ -2027,7 +2119,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.category": {
             "type": "string",
@@ -2055,7 +2147,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.category": {
             "type": "string",
@@ -2083,7 +2175,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.category": {
             "type": "string",
@@ -2112,7 +2204,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "db.system.name": {
             "type": "string",
@@ -2141,7 +2233,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "http.request.method": {
             "type": "string",
@@ -2170,7 +2262,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.category": {
             "type": "string",
@@ -2199,7 +2291,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "resource.render_blocking_status": {
             "type": "string",
@@ -2228,7 +2320,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "sentry.category": {
             "type": "string",
@@ -2256,7 +2348,7 @@ mod tests {
 
         normalize_client_address(&mut attributes, Some("192.168.1.1".parse().unwrap()));
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -2280,7 +2372,7 @@ mod tests {
 
         normalize_client_address(&mut attributes, None);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {}
         "#);
     }
@@ -2299,7 +2391,7 @@ mod tests {
 
         normalize_client_address(&mut attributes, Some("192.168.1.1".parse().unwrap()));
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -2332,7 +2424,7 @@ mod tests {
 
         normalize_client_address(&mut attributes, Some("2001:db8::1".parse().unwrap()));
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -2348,7 +2440,7 @@ mod tests {
 
         normalize_inject_client_address(&mut attributes, Some("192.168.1.1".parse().unwrap()));
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -2372,7 +2464,7 @@ mod tests {
 
         normalize_inject_client_address(&mut attributes, Some("192.168.1.1".parse().unwrap()));
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -2388,7 +2480,7 @@ mod tests {
 
         normalize_inject_client_address(&mut attributes, None);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {}
         "#);
     }
@@ -2399,7 +2491,7 @@ mod tests {
 
         normalize_inject_client_address(&mut attributes, Some("2001:db8::1".parse().unwrap()));
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "client.address": {
             "type": "string",
@@ -2424,7 +2516,7 @@ mod tests {
 
         normalize_span_category(&mut attributes);
 
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
+        assert_annotated_snapshot!(attributes, @r#"
         {
           "some.other.attribute": {
             "type": "string",