+++ relay-server/src/services/outcome/mod.rs
@@ -1,0 +1,5 @@
+mod aggregator;
+mod service;
+
+pub use self::aggregator::OutcomeAggregator;
+pub use self::service::*;