+++ relay-server/src/utils/multipart.rs
@@ -216,7 +216,7 @@ pub async fn read_bytes_into_item(
     if n_bytes > limit {
         let attachment_type = item.attachment_type().unwrap_or(AttachmentType::Attachment);
         let item_type = DiscardItemType::Attachment(DiscardAttachmentType::from(attachment_type));
-        let _ = item.reject_err(Outcome::Invalid(DiscardReason::TooLarge(item_type)));
+        let _ = item.reject_err(Outcome::Invalid(DiscardReason::ItemTooLarge(item_type)));
 
         Err(multer::Error::FieldSizeExceeded {
             limit: limit as u64,
@@ -253,7 +253,7 @@ pub async fn multipart_items(
                     if let multer::Error::FieldSizeExceeded { .. } = e {
                         let attachment_type = DiscardAttachmentType::from(attachment_type);
                         let item_type = DiscardItemType::Attachment(attachment_type);
-                        let discard_reason = DiscardReason::TooLarge(item_type);
+                        let discard_reason = DiscardReason::ItemTooLarge(item_type);
                         let _ = items.reject_err(Outcome::Invalid(discard_reason));
                     }
                 })?;
@@ -265,7 +265,8 @@ pub async fn multipart_items(
                 items.merge_with(item, |items, item, _| items.push(item));
                 if attachments_size > config.max_attachments_size() {
                     let item_type = DiscardItemType::Attachment(DiscardAttachmentType::Attachment);
-                    let _ = items.reject_err(Outcome::Invalid(DiscardReason::TooLarge(item_type)));
+                    let _ =
+                        items.reject_err(Outcome::Invalid(DiscardReason::ItemTooLarge(item_type)));
                     return Err(multer::Error::StreamSizeExceeded {
                         limit: config.max_attachments_size() as u64,
                     });